pyproject.toml
[project]
name = "workout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "streamlit>=1.56.0",
]
